% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.R
\name{generate_font}
\alias{generate_font}
\title{Generate a font}
\usage{
generate_font(input, output = gsub("sfd$", "ttf", input))
}
\arguments{
\item{input}{Input filename (e.g. a FontForge spline font database file).}

\item{output}{Font filename.}
}
\value{
The \code{output} filename invisible.
As a side effect creates a font file.
}
\description{
\code{generate_font()} generates a font file using FontForge.
}
\details{
I observe this doesn't always update the font file if not run in a clean R session.
}
