% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/font.r
\name{generate_sfd}
\alias{generate_sfd}
\title{Generate a FontForge sfd file}
\usage{
generate_sfd(
  font = c("square", "narrow"),
  output = paste0("dotaro_", font, ".sfd")
)
}
\arguments{
\item{font}{The desired Dotaro font i.e. "square" or "narrow".}

\item{output}{A string of desired FontForge spline font database file name.}
}
\value{
The \code{output} filename invisible.
As a side effect generates a FontForge spline font database file.
}
\description{
\code{generate_sfd()} generates a FontForge Spline Font Database (sfd) file for
the indicated Dotaro font.
}
